Equipment System Complete
Banzai Banzai!!!

It was by far the worst system yet.... I'm scared for my future.

But yeah, it works fine, you can change the equipment for each player, it changes their status, in the menu you can see the status change each equipment will bring and a coule other info, you can unequip it and it shows the one that is currenltly equipped.

It is saved between scenes so that should be everything for now.

This was a nightmare, good luck future me :D